Till Media management completed

+++ app/Http/Controllers/AdminUsersController.php
@@ -139,7 +139,9 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         $user =  User::findOrFail($id);
+	if(file_exists(public_path().'/user_dp/'.$user->photo->file)) {
 	unlink(public_path().'/user_dp/'.$user->photo->file);
+	}
 	$user->delete();
 	Session::flash('deleted_user', 'The user has been deleted');
 	return redirect('/admin/users');
